// client/src/Services/Models/ClassModel.ts
import { ClassSessionModel } from "./ClassSessionModel";
import { CourseModel } from "./CourseModel";
import { StatusMasterModel } from "./StatusMasterModel";
import { TutorModel } from "./TutorModels";

export interface ClassModel {
  class_pk?: number;
  pic?: string;
  course_pic?: string;
  class_desc?: string;
  course_pk?: number;
  course_desc?: string;
  course_duration?: number;
  room_pk?: number;
  room_desc?: string;
  class_type?: string;
  tutor_pk?: number;
  tutor_name?: string;
  tutor_pic?: string;
  start_date?: Date;
  start_time?: string;
  end_time?: string;
  session_count?: number;
  closed_sessions?: number;
  sts_pk?: string;
  remarks?: string;
  encoded_at?: Date;
  encoder_pk?: string;
  is_ongoing?: 1 | 0;
  class_sessions: Array<ClassSessionModel>;
  tutor_info?: TutorModel;
  status?: StatusMasterModel;
  ended_session?: number;
  course_info?: CourseModel;
  student_rating?: number;
  average_rating?: number;
  class_student_sts_pk?: string;
}